+++ src/twisted/trial/reporter.py
@@ -9,10 +9,7 @@ Defines classes that handle the results of tests.
 """
 
 
-import importlib
-import pathlib
 import sys
-import os
 import time
 import warnings
 import unittest as pyunit
@@ -25,30 +22,12 @@ from twisted.python import reflect, log
 from twisted.python.components import proxyForInterface
 from twisted.python.failure import Failure
 from twisted.python.util import untilConcludes
-from twisted.trial import itrial, util
+from . import itrial, util, _frames
 
 try:
     from subunit import TestProtocolClient  # type: ignore[import]
 except ImportError:
     TestProtocolClient = None
-
-
-def _getCvarFrame():
-    try:
-        from contextvars import Context
-    except ModuleNotFoundError:
-        return ("run", "defer")  # defer._NoContext.run
-
-    try:
-        _cvar_mod = Context.run.__module__
-    except AttributeError:
-        return None  # cpython 3.7 Context.run is a builtin
-
-    # pypy3.7 or https://pypi.org/project/contextvars
-    return ("run", pathlib.Path(importlib.import_module(_cvar_mod).__file__).stem)
-
-
-_cvarFrame = _getCvarFrame()
 
 
 def _makeTodo(value):
@@ -370,143 +349,6 @@ class _AdaptedReporter(TestResultDecorator):
         return self._originalReporter.stopTest(self.testAdapter(test))
 
 
-def _trimRunnerFrames(frames):
-    def frameInfo(f):
-        return f[0], os.path.splitext(os.path.basename(f[1]))[0]
-
-    syncCase = [("_run", "_synctest")]
-
-    if len(frames) < 1:
-        return frames
-
-    frame0 = frameInfo(frames[0])
-    if [frame0] == syncCase:
-        return frames[1:]
-
-    if len(frames) < 2:
-        return frames
-
-    asyncCase = [
-        ("_inlineCallbacks", "defer"),
-        ("_runCorofnWithWarningsSuppressed", "_asynctest"),
-    ]
-
-    frame1 = frameInfo(frames[1])
-    if [frame0, frame1] == asyncCase:
-        return frames[2:]
-
-    if len(frames) < 3:
-        return frames
-
-    asyncShimCvarCase = [
-        ("_inlineCallbacks", "defer"),
-        _cvarFrame,
-        ("_runCorofnWithWarningsSuppressed", "_asynctest"),
-    ]
-
-    frame2 = frameInfo(frames[2])
-    if [frame0, frame1, frame2] == asyncShimCvarCase:
-        return frames[3:]
-
-    if len(frames) < 4:
-        return frames
-
-    asyncFailureCase = [
-        ("_inlineCallbacks", "defer"),
-        ("throwExceptionIntoGenerator", "failure"),
-        ("_runCorofnWithWarningsSuppressed", "_asynctest"),
-        ("_runCallbacks", "defer"),
-    ]
-
-    frame2 = frameInfo(frames[2])
-    frame3 = frameInfo(frames[3])
-    if [frame0, frame1, frame2, frame3] == asyncFailureCase:
-        return frames[4:]
-
-    if len(frames) < 5:
-        return frames
-
-    asyncFailureShimCvarCase = [
-        ("_inlineCallbacks", "defer"),
-        _cvarFrame,
-        ("throwExceptionIntoGenerator", "failure"),
-        ("_runCorofnWithWarningsSuppressed", "_asynctest"),
-        ("_runCallbacks", "defer"),
-    ]
-
-    frame4 = frameInfo(frames[4])
-    if [frame0, frame1, frame2, frame3, frame4] == asyncFailureShimCvarCase:
-        return frames[5:]
-
-    return frames
-
-
-def _trimFrames(frames):
-    """
-    Trim frames to remove internal paths.
-
-    When a C{SynchronousTestCase} method fails synchronously, the stack
-    looks like this:
-     - [0]: C{SynchronousTestCase._run}
-     - [1:-2]: code in the test method which failed
-     - [-1]: C{_synctest.fail}
-
-    When a C{TestCase} method fails synchronously, the stack looks like
-    this:
-     - [0]: C{defer._inlineCallbacks}
-     - [1]: C{TestCase._runCorofnWithWarningsSuppressed}
-     - [2:-2]: code in the test method which failed
-     - [-1]: C{_synctest.fail}
-
-    When a method fails inside a C{Deferred} (i.e., when the test method
-    returns a C{Deferred}, and that C{Deferred}'s errback fires), the stack
-    captured inside the resulting C{Failure} looks like this:
-     - [0]: C{defer._inlineCallbacks}
-     - [1]: C{defer.throwExceptionIntoGenerator}
-     - [2]: C{TestCase._runCorofnWithWarningsSuppressed}
-     - [3]: C{defer._runCallbacks}
-     - [4:-2]: code in the testmethod which failed
-     - [-1]: C{_synctest.fail}
-
-    As a result, we want to trim those frames from the front,
-    and trim the [unittest.fail] from the end.
-
-    There is also another case, when the test method is badly defined and
-    contains extra arguments.
-
-    If it doesn't recognize one of these cases, it just returns the
-    original frames.
-
-    @param frames: The C{list} of frames from the test failure.
-
-    @return: The C{list} of frames to display.
-    """
-    newFrames = _trimRunnerFrames(list(frames))
-
-    if not newFrames:
-        # The method fails before getting called, probably an argument
-        # problem
-        return newFrames
-
-    last = newFrames[-1]
-    if (
-        last[0].startswith("fail")
-        and os.path.splitext(os.path.basename(last[1]))[0] == "_synctest"
-    ):
-        return newFrames[:-1]
-
-    return newFrames
-
-
-def _formatFailureTraceback(*, detail, fail):
-    if isinstance(fail, str):
-        return fail.rstrip() + "\n"
-    fail.frames, frames = _trimFrames(fail.frames), fail.frames
-    result = fail.getTraceback(detail=detail, elideFrameworkCode=True)
-    fail.frames = frames
-    return result
-
-
 @implementer(itrial.IReporter)
 class Reporter(TestResult):
     """
@@ -655,7 +497,7 @@ class Reporter(TestResult):
         )
 
     def _formatFailureTraceback(self, fail):
-        return _formatFailureTraceback(detail=self.tbformat, fail=fail)
+        return _frames._formatFailureTraceback(detail=self.tbformat, fail=fail)
 
     def _groupResults(self, results, formatter):
         """
