+++ src/twisted/trial/test/test_reporter.py
@@ -21,7 +21,7 @@ from unittest import TestCase as StdlibTestCase
 from twisted.python import log, reflect
 from twisted.python.failure import Failure
 from twisted.python.reflect import qual
-from twisted.trial import itrial, unittest, runner, reporter, util
+from twisted.trial import itrial, unittest, runner, reporter, util, _frames
 from twisted.trial.reporter import _ExitWrapper, UncleanWarningsReporterWrapper
 from twisted.trial.test import erroneous
 from twisted.trial.unittest import makeTodo, SkipTest, Todo
@@ -268,7 +268,7 @@ class TracebackHandlingTests(unittest.SynchronousTestCase):
         bads = result.failures + result.errors
         self.assertEqual(len(bads), 1)
         self.assertEqual(bads[0][0], test)
-        return reporter._trimFrames(bads[0][1].frames)
+        return _frames._trimFrames(bads[0][1].frames)
 
     def checkFrames(self, observedFrames, expectedFrames):
         for observed, expected in zip(observedFrames, expectedFrames):
@@ -304,10 +304,10 @@ class TracebackHandlingTests(unittest.SynchronousTestCase):
         self.checkFrames(frames, [("_later", "twisted/trial/test/erroneous")])
 
     def test_noFrames(self):
-        self.assertEqual([], reporter._trimFrames([]))
+        self.assertEqual([], _frames._trimFrames([]))
 
     def test_oneFrame(self):
-        self.assertEqual(["fake frame"], reporter._trimFrames(["fake frame"]))
+        self.assertEqual(["fake frame"], _frames._trimFrames(["fake frame"]))
 
     def test_exception(self):
         """
